use shapelen in padding

+++ examples/batch/batch.go
@@ -18,7 +18,7 @@ import (
 )
 
 var (
-	batch        = uint32(10)
+	batch        = 10
 	graph_url    = "http://data.dmlc.ml/models/imagenet/squeezenet/squeezenet_v1.0-symbol.json"
 	weights_url  = "http://data.dmlc.ml/models/imagenet/squeezenet/squeezenet_v1.0-0000.params"
 	features_url = "http://data.dmlc.ml/mxnet/models/imagenet/synset.txt"
@@ -83,7 +83,7 @@ func main() {
 		mxnet.Symbol(symbol),
 		mxnet.Weights(params),
 		mxnet.InputNode("data", []uint32{3, 224, 224}),
-		mxnet.BatchSize(batch),
+		mxnet.BatchSize(uint32(batch)),
 	)
 	if err != nil {
 		panic(err)
